requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
backtrader==1.9.78.123
backtrader-bybit==2.0.9
binance-futures-connector==4.0.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.1.1
fastapi-cli==0.0.4
fonttools==4.53.0
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pybit==5.7.0
pycryptodome==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
pyTelegramBotAPI==4.19.2
python-dateutil==2.9.0.post0
python-dotenv>=0.19.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
ta==0.11.0
termcolor==2.3.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvloop==0.19.0
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
yaspin==3.0.2